split domain package
